backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
requests==2.32.3
apscheduler==3.10.4
email-validator==2.2.0
python-jose==3.3.0
